requirements.txt
flask==2.3.3
streamlit==1.28.1
requests==2.31.0
python-dotenv==1.0.0
openai==1.3.0
